Rsync local files
Instead of symlinking /vagrant to /var/www/PROJECT/source. This
way the project user can own the files and write to them.

We might want to not include --delete on the rsync command, though,
since that'll remove any files added to the source tree that
aren't in /vagrant, like log files.

Branch: rsync-local-files