Update margarata git checkout more manually
The salt git.latest state just doesn't seem able to
cope with all the possible ways the remote git might
need to be updated or that we might have changed what
we want it to check out.

Switch to cloning the repo directly if it's not
there already, always doing a fetch to update our
view of the origin, then using a hard reset to
forcibly switch to the desired revision, branch, or
tag.

Branch: move_states_to_margarita