Get modernizr build working
We want to use modernizr in a project, and this
gulpfile appears to support building a custom
modernizr based on what use we are making of it.
So we tried it.

Unfortunately, this required a major refactoring
of gulpful.js, to make sure we had built modernizr.js
before we tried to compile our javascript.  On
the plus side, it's more gulpish now.

Branch: modernizr