Touch output of gulp compiles to update modtime
Gulp is setting the modtime on bundle.css to the modtime of index.less, even if other .less files that it includes are newer. And collectstatic only copies files if the source file's modtime is newer than an existing target file's modtime. So, Gulp might be recompiling the styles, but collectstatic won't copy them because they don't look newer than what's already there.

    To fix this, the gulpfile is modified to set the modtime of the result of each compilation to the current time.